% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_grouped_log.R
\name{as_grouped_log}
\alias{as_grouped_log}
\title{Group log by cluster}
\usage{
as_grouped_log(x, ...)
}
\arguments{
\item{x}{An object of type `trace_clustering` as returned by `cluster_traces()`.}

\item{...}{Additional parameters}
}
\value{
A grouped log on cluster variable
}
\description{
Group log by its respective trace cluster as calculated with `cluster_traces()`. It is possible to access
information on traces per cluster with `attr()`.
}
