% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_cluster.R
\name{augment_cluster}
\alias{augment_cluster}
\title{Augment original log with cluster ids}
\usage{
augment_cluster(trace_clustering)
}
\arguments{
\item{trace_clustering}{Result of `cluster_traces`.}
}
\value{
Log with added column `cluster`
}
\description{
Augment original log with cluster ids
}
