% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_traces.R
\name{cluster_traces}
\alias{cluster_traces}
\title{Cluster an event log based on its traces.}
\usage{
cluster_traces(
  log = eventdataR::patients,
  num_clusters = 3,
  target_fitness = 1,
  mcs = 0.6,
  window_size = 0.25,
  cluster_remaining_traces = F,
  algo = pm4py::discovery_inductive,
  distance_metric = stringdist(method = "osa"),
  ...
)
}
\arguments{
\item{log}{[log]: An object of type [`log`].}

\item{num_clusters}{[num]: A number of clusters.}

\item{target_fitness}{[num]: A target value to compare with the log fitness.}

\item{mcs}{[num]: A minimum cluster size to test whether the cluster is already big enough.}

\item{window_size}{[num]: A window size of the most frequent traces, i.e. traces with the highest absolute frequency. If `1`, all of the remaining traces are evaluated in the distance based method. If `0`, only one trace with the highest absolute frequency is chosen (frequency-based method).}

\item{cluster_remaining_traces}{[lgl] (default `FALSE`): Whether to cluster the remaining traces together. If `FALSE` (default), the remaining traces are assigned to that cluster, with which it has the best fitness. Otherwise, clusters the remaining traces together in a new additional cluster.}

\item{algo}{(default `pm4py::discovery_inductive`): A discovery algorithm.}

\item{distance_metric}{(default "stringdist"): A distance calculation method. By default ("stringdist"), `stringdist()` of a `stringdist` package is used. The literature suggest "mra".}

\item{...}{Additional parameters}
}
\value{
[list] A `list` object containing event logs filtered on trace id's, as a result of clustering algorithm.
}
\description{
Implementation of the active trace clustering (ActiTraC) algorithm.
}
\seealso{
This function is an implementation of the active trace clustering (ActiTraC) algorithm: J. De Weerdt, S. vanden Broucke, J. Vanthienen and B. Baesens, "Active Trace Clustering for Improved Process Discovery," in IEEE Transactions on Knowledge and Data Engineering, vol. 25, no. 12, pp. 2708-2720, Dec. 2013, doi: 10.1109/TKDE.2013.64.
}
