% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dict.R
\name{get_dict}
\alias{get_dict}
\title{Create a dictionary of activity id's}
\usage{
get_dict(log)
}
\arguments{
\item{log}{[`log`] An object of type [`log`].}
}
\value{
[list] A dictionary object of type `list`.
}
\description{
Creates a dictionary of unique activity id's.
}
