% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equivalence_classes.R
\name{get_equivalence_classes}
\alias{get_equivalence_classes}
\title{Get equivalence classes from a set of traces}
\usage{
get_equivalence_classes(tokenized_traces, dict)
}
\arguments{
\item{tokenized_traces}{[data.frame] A set of unique traces containing some additional metrics, a trace list and a tokenized trace list (tokens) as returned by `get_tokenized_traces()`.}

\item{dict}{A dictionary as returned by `get_dict()`.}
}
\value{
A set of equivalence classes.
}
\description{
Get equivalence classes from a set of traces based described in ...
}
