% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tokenized_traces.R
\name{get_tokenized_traces}
\alias{get_tokenized_traces}
\title{Get tokenized set of traces}
\usage{
get_tokenized_traces(log, dict)
}
\arguments{
\item{log}{[`log`] An object of type [`log`].}

\item{dict}{[list] A dictionary as returned by `get_dict()`.}
}
\value{
[data.frame] A collection of unique traces with additional information.
}
\description{
Get a set of unique traces containing some additional metrics, a trace list and a tokenized trace list (tokens).
}
