% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_cluster.R
\name{group_by_cluster}
\alias{group_by_cluster}
\title{Group by trace clustering}
\usage{
group_by_cluster(trace_clustering)
}
\arguments{
\item{trace_clustering}{Result of `cluster_traces`.}
}
\value{
Log with added column `cluster`, and grouped on that column.
}
\description{
Group by trace clustering
}
