% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trace_clustering.R
\name{print.trace_clustering}
\alias{print.trace_clustering}
\title{Print clustering}
\usage{
\method{print}{trace_clustering}(x, ...)
}
\arguments{
\item{x}{Clustering}

\item{...}{Not used.}
}
\value{
Prints the result of `cluster` function.
}
\description{
Print clustering
}
