% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cluster.R
\name{select_cluster}
\alias{select_cluster}
\title{Select a cluster}
\usage{
select_cluster(trace_clustering, cluster_id)
}
\arguments{
\item{trace_clustering}{Result of `cluster_traces`.}

\item{cluster_id}{cluster to select}
}
\value{
Log of selected cluster
}
\description{
Select a cluster
}
