% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_traces.R
\name{cluster_traces}
\alias{cluster_traces}
\title{Cluster an event log based on its traces.}
\usage{
cluster_traces(
  log,
  number_of_clusters = 3,
  target_fitness = 0.8,
  minimal_cluster_size = 0.6,
  window_size = 0.25,
  cluster_remaining_traces = F,
  discover_algorithm = pm4py::discover_inductive,
  distance_metric = stringdist(method = "osa"),
  ...
)
}
\arguments{
\item{log}{[log]: An object of type [`log`].}

\item{number_of_clusters}{[num]: A number of clusters.}

\item{target_fitness}{[num]: A target value to compare with the log fitness.}

\item{minimal_cluster_size}{[num]: A minimal cluster size. Defined as a percentage.}

\item{window_size}{[num]: A window size of the most frequent traces, i.e. traces with the highest absolute frequency. If `1`, all of the remaining traces are evaluated in the distance based method. If `0`, only one trace with the highest absolute frequency is chosen (frequency-based method).}

\item{cluster_remaining_traces}{[lgl] (default `FALSE`): Whether to cluster the remaining traces together. If `FALSE` (default), the remaining traces are assigned to that cluster, with which it has the best fitness. Otherwise, clusters the remaining traces together in a new additional cluster.}

\item{discover_algorithm}{(default `pm4py::discover_inductive`): A process discovery algorithm, returning a Petri Net}

\item{distance_metric}{(default "stringdist"): Distance calculation method.
- stringdist: `stringdist()` of a `stringdist` package is used.
- mra: Maximal repeat alphabet}

\item{...}{Additional parameters for discovery algorithm}
}
\value{
[list] A `list` object containing event logs filtered on trace id's, as a result of clustering algorithm.
}
\description{
Implementation of the active trace clustering (ActiTraC) algorithm.
}
\seealso{
This function is an implementation of the active trace clustering (ActiTraC) algorithm: J. De Weerdt, S. vanden Broucke, J. Vanthienen and B. Baesens, "Active Trace Clustering for Improved Process Discovery," in IEEE Transactions on Knowledge and Data Engineering, vol. 25, no. 12, pp. 2708-2720, Dec. 2013, doi: 10.1109/TKDE.2013.64.
}
